block part of frame, log images with kites

+++ vision/vision.py
@@ -28,29 +28,26 @@ def get_stream_url(url):
 
     return url
 
-def count_kites(url):
+def count_kites(url, extra_args=[]):
     CLS_KITE = 33
-
-    subprocess.run(["ffmpeg", "-i", url, "-ss", "2", "-frames:v", "1", "/tmp/frame.jpg"],
+    subprocess.run(["ffmpeg", "-i", url, "-ss", "2", *extra_args, "-frames:v", "1", "/tmp/frame.jpg"],
                 check=True, capture_output=True)
-    results = model("/tmp/frame.jpg")
-    os.remove("/tmp/frame.jpg")
+    results = model("/tmp/frame.jpg", classes=[CLS_KITE], imgsz=1024)
     detections = results[0].boxes
-    print(detections)
-
     kite_count = sum(1 for box in detections if box.cls == CLS_KITE)
 
-    if not DEBUG:
-        return kite_count, None
+    if kite_count > 0:
+        timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
+        debug_dir = 'vision_debug/'
+        os.makedirs(debug_dir, exist_ok=True)
+        os.rename("/tmp/frame.jpg", f"{debug_dir}input_{timestamp}_{kite_count}.jpg")
+        annotated_img = results[0].plot()
+        result_path = f"{debug_dir}result_{timestamp}_{kite_count}.jpg"
+        cv2.imwrite(result_path, annotated_img)
+        return kite_count, result_path
 
-    annotated_img = results[0].plot()
-    debug_dir = 'vision_debug/'
-    os.makedirs(debug_dir, exist_ok=True)
-    timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-    result_path = f"{debug_dir}result_{timestamp}.jpg"
-    cv2.imwrite(result_path, annotated_img)
-
-    return kite_count, result_path
+    os.remove("/tmp/frame.jpg")
+    return kite_count, None
 
 
 class CountBuffer:
@@ -67,22 +64,31 @@ class CountBuffer:
             return 0
         return sum(self.counts) / len(self.counts)
 
+def write_output(j):
+    with open(OUTPUT_FILE, 'w') as f:
+        json.dump(j, f)
+
+def log_output(s):
+    print(s)
+    with open(LOG_FILE, 'a') as f:
+        f.write(f"{datetime.datetime.now()}: {s}\n")
 
 def main():
+    # Langebaan cams
+    # "https://www.youtube.com/watch?v=Hhc6vesmrNk"
+    # "https://s75.ipcamlive.com/streams/4bxjaol4nvopmqhtb/stream.m3u8"
     SPOTS = [
         {
             "slug": "khaya",
             "url": "https://conjure.co.za/blouberg/hls/media.m3u8",
+            # Block out windsock that gets mistaken for a kite sometimes
+            "ffmpeg_extra_args": ["-vf", "drawbox=x=175:y=1015:w=200:h=100:color=black:t=fill"],
             "counter": CountBuffer()
         },
-        # {
-        #     "slug": "langebaan",
-        #     "counter": KiteCounter("https://www.youtube.com/watch?v=Hhc6vesmrNk")
-        #     "counter": KiteCounter("https://s75.ipcamlive.com/streams/4bxjaol4nvopmqhtb/stream.m3u8")
-        # },
         {
             "slug": "bigbay",
             "url": "https://live-sec.streamworks.video/oceaneye/oceaneye12.stream/chunks.m3u8",
+            "ffmpeg_extra_args": [],
             "counter": CountBuffer()
         }
     ]
@@ -93,38 +99,31 @@ def main():
         is_daytime = 7 <= now.hour <= 21
         if not is_daytime:
             print("Not daytime, skipping")
-            with open(OUTPUT_FILE, 'w') as f:
-                json.dump({
-                    "khaya": None,
-                    "langebaan": None,
-                    "bigbay": None
-                }, f)
+            write_output({
+                "khaya": None,
+                "langebaan": None,
+                "bigbay": None
+            })
             time.sleep(FETCH_SLEEP_SECS)
             continue
 
         for spot in SPOTS:
-            # Run synchronously
             count = 0
             try:
-                kite_count, _ = count_kites(get_stream_url(spot["url"]))
+                kite_count, _ = count_kites(get_stream_url(spot["url"]), spot["ffmpeg_extra_args"])
                 count = kite_count
             except Exception as e:
                 print(f"Error counting kites: {e}")
 
             spot["counter"].observe(count)
-            with open(LOG_FILE, 'a') as f:
-                f.write(f"{now}: [{spot['slug']}]: {count}\n")
+            log_output(f"[{spot['slug']}]: {count}")
 
         avg = {}
         for spot in SPOTS:
             avg[spot["slug"]] = spot["counter"].average()
 
-        print(avg)
-        with open(OUTPUT_FILE, 'w') as f:
-            json.dump(avg, f)
-
-        with open(LOG_FILE, 'a') as f:
-            f.write(f"{now}: [avg]: {json.dumps(avg)}\n")
+        write_output(avg)
+        log_output(f"[avg]: {json.dumps(avg)}")
 
         if not LOOP:
             break
